cli: Fixed Log file initialization
Updates: #205
